:truck: move initialize_main_data_points to tab1 as it is only used there
